Refactor types.rs: Move structs to appropriate modules with re-exports
- Create src/api/types.rs for API-related types (Pagination, ApiResponse, Error types)
- Move health check types to src/api/health.rs alongside existing health endpoints
- Update src/api/mod.rs to export new types module
- Maintain backward compatibility by re-exporting moved types from src/types.rs
- Keep core application types (AppState, DbPool, DbConn) in src/types.rs
- All tests pass, no breaking changes to public API
