+++ starter/src/api/mod.rs
@@ -1,3 +1,5 @@
 pub mod health;
+pub mod types;
 
 pub use health::*;
+pub use types::*;
